Code From JIT-SDP Research/requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==5.1.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pyzmq==25.0.0
scikit-learn==1.2.0
scipy==1.10.0
setuptools==65.6.3
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
wcwidth==0.2.6
wheel==0.37.1
